requirements.txt
psycopg==3.2
django==4.2.19
djangorestframework~=3.15.2
django-cors-headers==4.7.0
environs==14.1.1
pillow==11.1.0
